third_party/blink/web_tests/http/tests/security/xss-DENIED-frame-name-expected.txt

CONSOLE ERROR: Uncaught SecurityError: Blocked a frame with origin "http://localhost:8000" from accessing a frame with origin "http://127.0.0.1:8000". Protocols, domains, and ports must match.
  This test passes if it doesn't alert fail.
